Widen the pre-load buffer for stops to 0.7
We used `map.getBounds().pad(0.5)` to chose which stops to load to give
the illusion (for small pans and zooms) that all the stops were always
loaded. However following one zoom out the tops of some stop markers
tended to appear at the bottom of the map. This is presumably because
the stop had been excluded but its marker extended into the now-
visible region. Changing the buffering to `map.getBounds().pad(0.7)`
seems to resolve this.