Add `lng` and `lat` properties to stop data
Pending addition to the API, add `lng` and `lat` properties to
stop objects, duplicating the existing 'longitude' and 'latitude'
properties for consistency with processing elsewhere.

At the same time, refactor the rest of the code to use `stop_id`, `lng`
and `lat` in place of `atco_code`, `longitude` and `latitude`.