Add trailing slash to generated API URL
All API URLs strictly end in '/'. However Django automatically
redirects a URL without a trailing slash if it doesn't exist but the
corresponding one with a slash does, so in practice it doesn't always
matter if the trailing slash is omitted. However there's some indication
that such redirections may not always play well with CORS leading to
problems when making cross-origin requests. This edit corrects the URL
generated by the Bus Stop Chooser so that it includes the trailing slash.