provide an alternative on_click event
The default on_click event allows user to select single or multiple bus stops. Extend the functionality of the BusStopChooser by providing an alternative on_click event to the default one that will redirect the user to the bus stop web app endpoint. This can be controlled by the param option onclick_redirect.