Fix default click handler
An earlier change stopped the internal stop click handler from working,
making it impossible to actually select or deselect stops. This edit
reverts the broken change and renames a variable ('selected' -->
'already_selected') to reduce the potential for the confusion that lead
to the breakage.

This edit also adds a demo of the new on_click_stop_callback parameter
to demo.html.